Working on Enhancement 10 bug fixed

+++ app/Http/Controllers/consultantsController.php
@@ -101,7 +101,7 @@ class consultantsController extends Controller
             'secondary_skill'=>$request->secondary_skill,
             'secondary_skill_experience'=>$request->secondary_skill_experience,
             'resume'=>$fileName,
-            'is_active' => $request->is_active,
+            // 'is_active' => $request->is_active,
             'new_job_report_time'=>$request->new_job_report_time
         ]);
 
@@ -123,14 +123,19 @@ class consultantsController extends Controller
             ], 404);
         }
 
-        $Constant->delete();
+        if($Constant->is_active == 1){
+            return response()->json([
+                'message' => 'Can not delete active Consultant',
+                'error' =>100
+            ]);
+        }
 
+        $Constant->delete();
         return response()->json([
             'message' => 'Details deleted successfully',
             'success' => 200,
-            'consultantas_id' => $id
+            'consultantas_id' => $Constant
         ]);
-
     }
 
     public function statusConsultantsDetails(Request $request){
